participant-integration-api: Rename a class. Again.
